AI & Shotgun
The AI movement is complete. I tried out a new movement type on the
shotgunner to see if it works. Kamikaze now flashes before exploding.
Dodger has a weapon but fires player bullets (have not created enemy
projectiles yet. A protoype shotgun is currently in progress with a
dynamic shell-count in mind.